backend/requirements.txt
FastAPI
uvicorn
pydantic
aio_pika
requests
python-dotenv
pytest
pytest-asyncio